requirements.txt
# Azure VNET Management API - Requirements

# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.0
pydantic-settings==2.5.2

# Azure SDK
azure-identity==1.18.0
azure-mgmt-network==27.0.0
azure-mgmt-resource==23.1.1
azure-data-tables==12.7.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Environment Management
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2